backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
aiohttp
python-jose[cryptography]